TensorVis: compile fixes for VS build

+++ tensorvis/tensorvisbase/src/datastructures/tensorfield2d.cpp
@@ -550,7 +550,7 @@ void TensorField2D::computeEigenValuesAndEigenVectors() {
     minorEigenVectors_.resize(size_);
 
 #pragma omp parallel for
-    for (size_t i = 0; i < tensors_.size(); i++) {
+    for (int i = 0; i < static_cast<int>(tensors_.size()); i++) {
         auto tensor = tensors_[i];
 
         auto eigenValuesAndEigenVectors = func(tensor);
